Version 12F of the ESP8266
This is based on the 12E version by thedub2001.
However, its breadboard view was completely reworked to match
the actual appearance. Also AFAICT the additional pins of the
12E version are in the wrong order.
Finally, I have also changed the grouping of the pins in the
schematic to make a little bit more sense.
However, I had to change the text in the respective SVG to paths
because Inkscape and Fritzing were not able to agree on the
correct appearance of the font.